django/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
django-mysql==4.5.0
djangorestframework==3.13.1
idna==3.3
mysqlclient==2.1.0
pika==1.2.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
